added can only updata/delete cat

+++ models/catModel.js
@@ -19,11 +19,12 @@ const getCat = async (catId, next) => {
   }
 };
 
-const getAllCats = async (next) => {
+const getAllCats = async (next, req) => {
   try {
     // TODO: do the LEFT (or INNER) JOIN to get owner's name as ownername (from wop_user table).
     const [rows] = await promisePool.query(
-        'SELECT cat_id, owner, wop_cat.name AS name, weight, birthdate, filename, wop_user.name AS ownername FROM wop_cat INNER JOIN wop_user ON owner = user_id');
+        'SELECT cat_id, owner, wop_cat.name AS name, weight, birthdate, filename, wop_user.name AS ownername FROM wop_cat INNER JOIN wop_user ON owner = user_id Where owner=?',
+        [req.user.user_id]);
     return rows;
   } catch (e) {
     console.error('error', e.message);
@@ -32,7 +33,7 @@ const getAllCats = async (next) => {
   }
 };
 
-const insertCat = async (cat,next) => {
+const insertCat = async (cat, next) => {
   try {
     // TODO add filename
     const [rows] = await promisePool.execute(
@@ -47,11 +48,18 @@ const insertCat = async (cat,next) => {
   }
 };
 
-const deleteCat = async (catId,next) => {
+const deleteCat = async (catId, next, user) => {
   try {
-    const [rows] = await promisePool.execute(
-        'DELETE FROM wop_cat WHERE cat_id=?', [catId]);
+    let rows;
+    if (user.role === 1) {
+      [rows] = await promisePool.execute(
+          'DELETE FROM wop_cat WHERE cat_id=? and owner=?', [catId, user]);
+    } else {
+       [rows] = await promisePool.execute(
+          'DELETE FROM wop_cat WHERE cat_id=?', [catId]);
+    }
     console.log('model delete cat', rows);
+    return rows.affectedRows === 1;
     return true;
   } catch (e) {
     console.error('error', e.message);
@@ -60,11 +68,11 @@ const deleteCat = async (catId,next) => {
   }
 };
 
-const updateCat = async (cat,next) => {
+const updateCat = async (cat, next, user) => {
   try {
     const [rows] = await promisePool.execute(
-        `UPDATE wop_cat SET name=?, weight=?, owner=?, birthdate=?  WHERE cat_id=?`,
-        [cat.name, cat.weight, cat.owner, cat.birthdate, cat.id]);
+        `UPDATE wop_cat SET name=?, weight=?, owner=?, birthdate=?  WHERE cat_id=? AND owner=?`,
+        [cat.name, cat.weight, cat.owner, cat.birthdate, cat.id, user]);
     console.log('model update cat', rows);
     return rows.affectedRows === 1;
   } catch (e) {
